+++ src/application/dto/user/UserResponse.ts
@@ -1,0 +1,5 @@
+export interface UserResponse {
+    id: string;
+    name: string;
+    email: string;
+  }
