+++ src/application/dto/user/UserUpdateDto.ts
@@ -1,0 +1,6 @@
+export interface UserUpdateDto {
+    name?: string;
+    email?: string;
+    password?: string;
+  }
+
